2011-09-12 Roman - Archive: Send time label is made for archive's values table for display local time on remote station. - !!:FIX:Archive.FSArch: Store "double" value into LE format on ARM is fixed.     Use function TSYS::getUnalign32() for get dwords to fast calc value offset algorithm to prevent errors on ARM. - FIX:DAQ.OPC_UA: Input request counter is fixed for OPC UA protocol. - FIX:Transport.SSL, Transport.Serial: Return negative result is prevented for output transport messIO() function.     Take place on multiply connections to serial devices.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@1553 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/moduls/daq/OPC_UA/mod_prt.h
@@ -384,7 +384,7 @@ class OPCEndPoint : public TCntrNode, public TConfig
 	bool	&mAEn, mEn;
 	string	mDB;
 
-	float	cntReq;
+	uint64_t		cntReq;
 	vector<SecuritySetting>	mSec;
 	vector<OPCSess>		mSess;
